+++ packages/ubcli/lib/createStore.js
@@ -101,4 +101,5 @@ module.exports = function createStore (options) {
   selectedStores.forEach(createOneStore)
 }
 
-module.exports.shortDoc = 'Create internal BLOB store structure (folders) for specifies FileSystem store'
+module.exports.shortDoc = `Create internal BLOB store structure (folders) for
+ \t\t\ta specifies FileSystem store`